LETS GO

+++ main/views.py
@@ -10,6 +10,7 @@ from .models import Post
 
 def index(request):
     posts = Post.objects.all()
+    posts = posts.order_by("-published_at")
     return render(request, "main/index.html", {"posts": posts})
 
 
@@ -17,9 +18,10 @@ def get_post(request):
     loader = instaloader.Instaloader()
     profile = Profile.from_username(loader.context, "popular.front")
     posts = profile.get_posts()
+    i = 0
     for post in posts:
         print(f"Downloading post: {post.shortcode}")
-        if not Post.objects.filter(code=post.shortcode).exists():
+        if not Post.objects.filter(code=post.shortcode).exists() and i not in [0, 1, 2]:
             if post.is_video:
                 image = f"{post.shortcode}.mp4"
                 with open(f"static/{image}", "wb") as f:
@@ -28,5 +30,14 @@ def get_post(request):
                 image = f"{post.shortcode}.jpg"
                 with open(f"static/{image}", "wb") as f:
                     f.write(requests.get(post.url).content)
-            Post.objects.create(code=post.shortcode, content=post.caption, image=image)
+            Post.objects.create(
+                code=post.shortcode,
+                content=post.caption,
+                image=image,
+                published_at=post.date,
+            )
+
+        if i == 13:
+            break
+        i += 1
     return HttpResponse("Posts saved")
